Fix clang warnings for android builds (#6141)
https://github.com/youtube/cobalt/pull/6139 will actually remove the
clang warning suppression flags after this is submitted.

Bug: 374302273
Bug: 378934120

original-hexsha: 4f4cd5752ad884352d2ed528dea8e1f1d05ec380

+++ starboard/nplb/posix_compliance/posix_socket_resolve_test.cc
@@ -42,9 +42,11 @@ TEST(PosixSocketResolveTest, SunnyDay) {
     if (ai_addr == nullptr && i->ai_addr != nullptr) {
       ai_addr = reinterpret_cast<sockaddr_in*>(i->ai_addr);
       if (i->ai_family == AF_INET) {
-        EXPECT_EQ(i->ai_addrlen, sizeof(struct sockaddr_in));
+        EXPECT_EQ(static_cast<size_t>(i->ai_addrlen),
+                  sizeof(struct sockaddr_in));
       } else if (i->ai_family == AF_INET6) {
-        EXPECT_EQ(i->ai_addrlen, sizeof(struct sockaddr_in6));
+        EXPECT_EQ(static_cast<size_t>(i->ai_addrlen),
+                  sizeof(struct sockaddr_in6));
       }
       break;
     }
